requirements.txt
click>=8.1.0
pyyaml>=6.0
requests>=2.31.0

# Monitoring dependencies
psutil>=5.9.0
psycopg2-binary>=2.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
